+++ src/main/java/callofcactus/entities/pickups/AmmoPickup.java
@@ -1,6 +1,6 @@
 package callofcactus.entities.pickups;
 
-import callofcactus.Game;
+import callofcactus.IGame;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.math.Vector2;
 
@@ -18,7 +18,7 @@ public class AmmoPickup extends Pickup {
 	 * @param spriteWidth   : The width of the pickup
 	 * @param spriteHeight  : The height of the pickup
 	 */
-	public AmmoPickup(Game game, Vector2 location, Texture spriteTexture, int spriteWidth, int spriteHeight) {
+	public AmmoPickup(IGame game, Vector2 location, Texture spriteTexture, int spriteWidth, int spriteHeight) {
 		super(game, location, spriteTexture, spriteWidth, spriteHeight);
 	}
