+++ src/main/java/callofcactus/menu/MainMenu.java
@@ -1,8 +1,8 @@
 package callofcactus.menu;
 
 import callofcactus.BackgroundRenderer;
-import callofcactus.Game;
 import callofcactus.GameInitializer;
+import callofcactus.IGame;
 import callofcactus.io.DatabaseManager;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
@@ -27,7 +27,7 @@ import java.util.List;
 public class MainMenu implements Screen {
 
 	private Stage stage;
-	private List<Game> games;
+	private List<IGame> games;
 	private GameInitializer gameInitializer;
 	private SpriteBatch batch;
 	//GUI fields
@@ -238,7 +238,7 @@ public class MainMenu implements Screen {
 	 *
 	 * @return the list of all the current games
 	 */
-	public List<Game> getAllGames() {
+	public List<IGame> getAllGames() {
 		// TODO - implement MainMenu.getAllLobbies
 		throw new UnsupportedOperationException();
 	}
