--- src/main/java/testClasses/GameMockup.java
@@ -1,23 +1,0 @@
-package testClasses;
-
-import java.rmi.RemoteException;
-
-/**
- * Created by Teun on 19-10-2015.
- */
-public class GameMockup extends callofcactus.Game {
-	public GameMockup() throws RemoteException {
-
-		super();
-	}
-
-	@Override
-	public void playRandomHitSound() {
-
-	}
-
-	@Override
-	public void playRandomBulletSound() {
-
-	}
-}
